+++ src/main/java/adapters/primary/CertificateController.java
@@ -3,8 +3,8 @@ package adapters.primary;
 import application.dtos.CertificateDTO;
 import application.use_cases.ValidateCertificateApplicationService;
 import domain.entities.CertificateValidationResult;
-import domain.exceptions.CertificateValidationException;
-import domain.exceptions.TrustChainRetrievalException;
+import application.exceptions.CertificateValidationException;
+import application.exceptions.TrustChainRetrievalException;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
@@ -29,11 +29,10 @@ public class CertificateController {
             @ApiResponse(code = 400, message = "Certificate validation error"),
             @ApiResponse(code = 500, message = "Trust chain retrieval error")
     })
-    public ResponseEntity<CertificateDTO> validateCertificate(@RequestBody CertificateDTO certificateDTO) {
+    public ResponseEntity<CertificateValidationResult> validateCertificate(@RequestBody CertificateDTO certificateDTO) {
         try {
             CertificateValidationResult validationResult = validateCertificateApplicationService.validateCertificate(certificateDTO);
-            CertificateDTO resultDTO = validateCertificateApplicationService.convertToDTO(validationResult);
-            return ResponseEntity.ok(resultDTO);
+            return ResponseEntity.ok(validationResult);
         } catch (CertificateValidationException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Certificate validation error", e);
         } catch (TrustChainRetrievalException e) {
@@ -41,23 +40,3 @@ public class CertificateController {
         }
     }
 }
-
-// TODO: Extract the following exceptions to their respective files in the package 'domain.exceptions'
-
-package domain.exceptions;
-
-public class CertificateValidationException extends Exception {
-    public CertificateValidationException(String message) {
-        super(message);
-    }
-}
-
-package domain.exceptions;
-
-public class TrustChainRetrievalException extends Exception {
-    public TrustChainRetrievalException(String message) {
-        super(message);
-    }
-}
-
-// TODO: Implement the method 'convertToDTO' in ValidateCertificateApplicationService that converts CertificateValidationResult to CertificateDTO
