+++ src/main/java/application/dtos/CertificateDTO.java
@@ -30,43 +30,3 @@ public class CertificateDTO {
     private String signature;
     private String serialNumber;
 }
-
-// TODO: Extract the following classes to their respective files in the package 'application.serializers'
-
-package application.serializers;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-class CustomLocalDateSerializer extends JsonSerializer<LocalDate> {
-
-    private static final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE;
-
-    @Override
-    public void serialize(LocalDate value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
-        gen.writeString(formatter.format(value));
-    }
-}
-
-package application.serializers;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-class CustomLocalDateDeserializer extends JsonDeserializer<LocalDate> {
-
-    private static final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE;
-
-    @Override
-    public LocalDate deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
-        return LocalDate.parse(p.getValueAsString(), formatter);
-    }
-}
