+++ src/main/java/application/exceptions/TrustChainEncodingException.java
@@ -1,0 +1,7 @@
+package application.exceptions;
+
+public class TrustChainEncodingException extends RuntimeException {
+    public TrustChainEncodingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
