+++ src/main/java/application/serializers/CustomLocalDateDeserializer.java
@@ -1,0 +1,18 @@
+package application.serializers;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import java.io.IOException;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class CustomLocalDateDeserializer extends JsonDeserializer<LocalDate> {
+
+    private static final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE;
+
+    @Override
+    public LocalDate deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
+        return LocalDate.parse(p.getValueAsString(), formatter);
+    }
+}
