+++ src/main/java/domain/entities/CertificateEntity.java
@@ -1,6 +1,8 @@
 package domain.entities;
 
 import java.security.PublicKey;
+import java.security.cert.CertificateEncodingException;
+import java.security.cert.X509Certificate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -14,8 +16,6 @@ import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
 import org.bouncycastle.openssl.PEMParser;
 import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo;
 import java.io.StringReader;
-import java.security.cert.CertificateEncodingException;
-import java.security.cert.X509Certificate;
 
 @Getter
 @Setter
@@ -46,13 +46,12 @@ public class CertificateEntity {
         return certificateValidationService.validateCertificate(this.toX509CertificateHolder());
     }
 
-    public List<CertificateEntity> getTrustChain() throws TrustChainRetrievalException {
+    public List<X509CertificateHolder> getTrustChain() throws TrustChainRetrievalException {
         return certificateValidationService.constructTrustChain(this.toX509CertificateHolder());
     }
 
     public void revoke() {
         this.isRevoked = true;
-        // TODO: Implement logic to persist the revocation status change if necessary
     }
 
     public X509CertificateHolder toX509CertificateHolder() throws CertificateEncodingException {
@@ -62,15 +61,12 @@ public class CertificateEntity {
                 X509Certificate certificate = (X509Certificate) parsedObj;
                 return new X509CertificateHolder(certificate.getEncoded());
             } else if (parsedObj instanceof SubjectPublicKeyInfo) {
-                return new X509CertificateHolder(new JcaPEMKeyConverter().getPublicKey((SubjectPublicKeyInfo) parsedObj).getEncoded());
+                PublicKey publicKey = new JcaPEMKeyConverter().getPublicKey((SubjectPublicKeyInfo) parsedObj);
+                return new X509CertificateHolder(publicKey.getEncoded());
             }
             throw new IllegalArgumentException("Unable to parse certificate data.");
         } catch (Exception e) {
             throw new CertificateEncodingException("Failed to convert CertificateEntity to X509CertificateHolder", e);
         }
     }
-
-    // TODO: Add repository interface or domain service to handle persistence operations such as saving the updated CertificateEntity.
-
-    // TODO: Define CertificateEntityRepository interface with necessary methods like save for persisting CertificateEntity changes.
 }
