+++ src/main/java/domain/entities/CertificateValidationResult.java
@@ -10,20 +10,4 @@ public class CertificateValidationResult {
 
     private boolean validationStatus;
     private List<String> trustChain;
-
-    public boolean getValidationStatus() {
-        return validationStatus;
-    }
-
-    public void setValidationStatus(boolean validationStatus) {
-        this.validationStatus = validationStatus;
-    }
-
-    public List<String> getTrustChain() {
-        return trustChain;
-    }
-
-    public void setTrustChain(List<String> trustChain) {
-        this.trustChain = trustChain;
-    }
 }
