+++ src/main/java/domain/ports/outgoing/HttpClientPort.java
@@ -1,0 +1,9 @@
+package domain.ports.outgoing;
+
+import java.io.IOException;
+import org.apache.http.HttpResponse;
+import org.apache.http.client.methods.HttpUriRequest;
+
+public interface HttpClientPort {
+    HttpResponse sendRequest(HttpUriRequest request) throws IOException;
+}
