+++ src/main/java/domain/services/CertificateValidationDomainService.java
@@ -3,6 +3,8 @@ package domain.services;
 import domain.ports.outgoing.HttpClientPort;
 import lombok.RequiredArgsConstructor;
 import org.bouncycastle.cert.X509CertificateHolder;
+import application.exceptions.CertificateValidationException;
+import application.exceptions.TrustChainRetrievalException;
 import java.security.cert.CertificateException;
 import java.util.List;
 
@@ -27,42 +29,4 @@ public class CertificateValidationDomainService {
             throw new TrustChainRetrievalException("Unable to construct trust chain", e);
         }
     }
-
-    // Custom exceptions
-    public static class CertificateValidationException extends CertificateException {
-        public CertificateValidationException(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
-
-    public static class TrustChainRetrievalException extends CertificateException {
-        public TrustChainRetrievalException(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
 }
-
-// Interfaces
-package domain.services;
-
-import org.bouncycastle.cert.X509CertificateHolder;
-import java.util.List;
-
-public interface CertificateValidator {
-    boolean validate(X509CertificateHolder certificateHolder) throws CertificateValidationException;
-}
-
-package domain.services;
-
-import org.bouncycastle.cert.X509CertificateHolder;
-import java.util.List;
-
-public interface TrustChainBuilder {
-    List<X509CertificateHolder> build(X509CertificateHolder certificateHolder) throws TrustChainRetrievalException;
-}
-
-// TODO: Extract the following interface to its respective package and class name
-// package domain.ports.outgoing;
-// public interface HttpClientPort {
-//     HttpResponse sendRequest(HttpUriRequest request) throws IOException;
-// }
