+++ src/main/java/domain/validation/ValidationResult.java
@@ -1,0 +1,13 @@
+package domain.validation;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ValidationResult {
+
+    private final boolean valid;
+    private final String errorMessage;
+
+}
