+++ src/main/java/infrastructure/external_services/CertificateValidationAdapter.java
@@ -5,12 +5,11 @@ import domain.ports.CertificateValidationPort;
 import domain.validation.ValidationResult;
 import java.util.List;
 import java.util.stream.Collectors;
-import infrastructure.external_services.AIAExtensionServiceClient;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 import application.exceptions.CertificateValidationException;
 import application.exceptions.TrustChainRetrievalException;
 import application.exceptions.TrustChainEncodingException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
@@ -25,7 +24,7 @@ public class CertificateValidationAdapter implements CertificateValidationPort {
         }
         try {
             boolean valid = certificateEntity.validate();
-            String message = valid ? "Certificate is valid." : certificateEntity.getValidationErrorMessage();
+            String message = valid ? "Certificate is valid." : "Certificate is invalid.";
             return new ValidationResult(valid, message);
         } catch (Exception e) {
             throw new CertificateValidationException("Validation failed", e);
@@ -53,19 +52,5 @@ public class CertificateValidationAdapter implements CertificateValidationPort {
         } catch (Exception e) {
             throw new TrustChainRetrievalException("Retrieval of trust chain failed", e);
         }
-    }
-
-    // TODO: Implement logic for interacting with the AIAExtensionServiceClient to fetch parent certificates.
-    // TODO: Implement error handling and retry logic for network-related issues when interacting with AIAExtensionServiceClient.
-    // TODO: Add repository interface or domain service to handle persistence operations such as saving the updated CertificateEntity.
-    // TODO: Define CertificateEntityRepository interface with necessary methods like save for persisting CertificateEntity changes.
-
-    // TODO: Define TrustChainEncodingException class in the application.exceptions package.
-}
-
-// TODO: Extract the following class to its own file in the application.exceptions package.
-class TrustChainEncodingException extends RuntimeException {
-    public TrustChainEncodingException(String message, Throwable cause) {
-        super(message, cause);
     }
 }
