+++ src/main/java/infrastructure/external_services/IAIAExtensionServiceClient.java
@@ -1,0 +1,9 @@
+package infrastructure.external_services;
+
+import java.io.IOException;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+
+public interface IAIAExtensionServiceClient {
+    X509Certificate fetchParentCertificate(String aiaExtensionUrl) throws IOException, CertificateException;
+}
